requirements.txt
paho-mqtt
psutil